pyproject.toml
[project]
name = "test-pytesseract"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "gradio>=4.0.0",
    "opencv-python>=4.8.0",
    "pandas>=2.0.0",
]
